Created separate package for exception handling, modified the frontend to match the new JSON format.
